rename optimizely client file
